fix: address code review findings — 5 issues from 4 reviewers
1. Fix ServiceContainer.resolve() caching bug (use cache.has instead of !== undefined)
2. Move MetricsResult from metrics/ to core/types.ts (eliminate upward dependency)
3. Move ResilienceAttempt from proxy/orchestration/resilience.ts to core/types.ts
4. Move UPSTREAM_SUCCESS, SKIP_DOWNSTREAM, filterHeaders from types.ts to constants.ts
5. Deduplicate container.resolve() calls in proxy-handler.ts failover loop

All 580 tests pass, zero compilation errors.
