+++ src/core/constants.ts
@@ -24,3 +24,24 @@ export function getProxyApiType(url: string): string | null {
 }
 
 export const MS_PER_SECOND = 1000;
+
+// 上游成功状态码
+export const UPSTREAM_SUCCESS = 200;
+
+/** 过滤掉不应转发给下游的 hop-by-hop headers */
+const SKIP_DOWNSTREAM = new Set([
+  "content-length",
+  "transfer-encoding",
+  "connection",
+  "keep-alive",
+]);
+
+/** 过滤掉不应转发给下游的 hop-by-hop headers */
+export function filterHeaders(raw: import("./types.js").RawHeaders): Record<string, string> {
+  const out: Record<string, string> = {};
+  for (const [key, value] of Object.entries(raw)) {
+    if (value == null || SKIP_DOWNSTREAM.has(key.toLowerCase())) continue;
+    out[key] = Array.isArray(value) ? value.join(", ") : value;
+  }
+  return out;
+}
