+++ src/proxy/transport/stream.ts
@@ -2,8 +2,7 @@ import { PassThrough } from "stream";
 import type { FastifyReply } from "fastify";
 import { UPSTREAM_SUCCESS, filterHeaders } from "../types.js";
 import { buildUpstreamUrl } from "../proxy-core.js";
-import type { RawHeaders, StreamState, TransportResult } from "../types.js";
-import type { MetricsResult } from "../../metrics/metrics-extractor.js";
+import type { RawHeaders, StreamState, TransportResult, MetricsResult } from "../types.js";
 import type { SSEMetricsTransform } from "../../metrics/sse-metrics-transform.js";
 import type { StreamLoopGuard } from "../loop-prevention/stream-loop-guard.js";
 import {
